Update dto classes

+++ src/main/java/ru/digitalmagicians/ebito/dto/FullAds.java
@@ -4,13 +4,13 @@ import lombok.Data;
 
 @Data
 public class FullAds {
-    private int pk;
+    private Integer pk;
     private String authorFirstName;
     private String authorLastName;
     private String description;
     private String email;
     private String image;
     private String phone;
-    private int price;
+    private Double price;
     private String title;
 }
